fix: fixing a bug that handles slug generation for existing files
Added logic to check an if the filename is the same as the slug, if it
is the filename will remain the same, if it's changed it will create a
new unique slug for the file.

This correctly makes it such that saving a file with the original name
doesn't add and remove _1 in sequential saves.